+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-# dev
+# Initial Commit
 
-- Init commit
+This commit introduces two utilities:
+
+- "scarchesh5ad2refbyscanvi": Trains a reference model using scArches.
+- "scarches2predictbyscanvi": Predicts cell types using the trained reference model.
+